Move SpecFixture type into its own file
